window confirm functionality added

+++ src/Pages/Card.jsx
@@ -9,16 +9,24 @@ export default function Card({ notedata, handleRetriveData, setEditCheck, editCh
 
     async function handleDeleteOne(item) {
 
-        try {
-            console.log(item.id);
-            //  const deleteOne = await axios.delete(`http://127.0.0.1:8000/api/delete/${item.id}/`);
-            const deleteOne = await axios.delete(`https://notesappapi-m3nt.onrender.com/api/delete/${item.id}/`);
-            handleRetriveData();
-            alert(`Success! \n \n ${item.title} Data deleted successfully`);
-        } catch (error) {
-            console.log(error);
+        const dialogvar = window.confirm(`Do you want to delete this ${item.title}?`)
 
+        if (dialogvar) {
+            try {
+                console.log(item.id);
+                //  const deleteOne = await axios.delete(`http://127.0.0.1:8000/api/delete/${item.id}/`);
+                const deleteOne = await axios.delete(`https://notesappapi-m3nt.onrender.com/api/delete/${item.id}/`);
+                handleRetriveData();
+                alert(`Success! \n \n ${item.title} Data deleted successfully`);
+            } catch (error) {
+                console.log(error);
+
+            }
+        } else {
+                console.log('Delete canceled');
+                
         }
+
 
     }
